# visol/design/prototype/API.yaml
openapi: 3.0.0
info:
  title: ViSOL
  description: >-
    Visual Interactive Optimization Tool.
  contact: { }
  version: 1.0.0
servers:
  - url: 'https://visol1.paas.hosted-by-previder.com/api/v1'
    variables: { }
paths:
  /token/obtain:
    post:
      tags:
        - Token
      summary: Obtain a token
      operationId: obtainToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  $ref: '#/components/schemas/Email'
                password:
                  type: string
                  format: password
                  pattern: '^(?:.|\s)*$'
                  maxLength: 1073741824
                  example: Password123
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    format: jwt
                    example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2NTY0NjAzMTEsImV4cCI6MTY1NjQ2MDM0MSwibmJmIjoxNjU2NDYwMTMxLCJ1c2VyIjoiamFuZXQ2OEB0b3JtYWlscy5jb20ifQ.Ap9vlr4sJ45Edd90vNc6K0RCYAJs1fPfIcnJWo098I1TMgARBM7yP4paf6nK1zZs4w7VXhxKY_kx9WAAu_8oIA
                additionalProperties: false
                required: [token]
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /token/verify:
    get:
      tags:
        - Token
      summary: Verify a token
      operationId: verifyToken
      responses:
        '200':
          description: 200 OK
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  '/vessels/{vessel_id}':
    parameters:
      - name: vessel_id
        in: path
        description: ID of the Vessel.
        required: true
        schema:
          $ref: '#/components/schemas/VesselID'
    get:
      tags:
        - Vessels
      summary: Retrieve a Vessel
      operationId: retrieveVessel
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vessel'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags:
        - Vessels
      summary: Replace a Vessel
      operationId: replaceVessel
      parameters:
        - in: header
          name: Reason
          schema:
            $ref: '#/components/schemas/Reason'
          required: false
          description: The reason for which the Vessel was created.
          example: Additional cargo (#2524)
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vessel'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vessel'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags:
        - Vessels
      summary: Delete a Vessel
      operationId: deleteVessel
      parameters:
        - in: header
          name: Reason
          schema:
            $ref: '#/components/schemas/Reason'
          required: false
          description: The reason for which the Vessel was deleted.
          example: Will no longer arrive in this port
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/vessels/{vessel_id}/schedule':
    parameters:
      - name: vessel_id
        in: path
        description: ID of the Vessel.
        required: true
        schema:
          $ref: '#/components/schemas/VesselID'
    get:
      tags:
        - Vessels
      summary: Retrieve a Vessel's Schedule
      operationId: retrieveVesselSchedule
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags:
        - Vessels
      summary: Replace or Create a Vessel's Schedule
      operationId: replaceOrCreateVesselSchedule
      parameters:
        - in: header
          name: Reason
          schema:
            $ref: '#/components/schemas/Reason'
          required: false
          description: The reason for which the Schedule was replaced/updated.
          example: Automatic scheduling job
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '201':
          description: ''
          headers:
            Location:
              schema:
                type: string
                pattern: '^\/vessels\/[0-9]{1,10}\/schedule$'
                maxLength: 28
                example: /vessels/1/schedule
              description: The URL of the newly created Schedule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags:
        - Vessels
      summary: Delete a Vessel's Schedule
      operationId: deleteVesselSchedule
      parameters:
        - in: header
          name: Reason
          schema:
            $ref: '#/components/schemas/Reason'
          required: false
          description: The reason for which the Schedule was deleted.
          example: Delayed for an indeterminate amount of time
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  '/changes/undo':
    post:
      tags:
        - Changes
      summary: Undo the last change
      operationId: undoLastChange
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/changes/redo':
    post:
      tags:
        - Changes
      summary: Redo the last change
      operationId: redoLastChange
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
security:
  - bearerAuth: [ ]
components:
  schemas:
    Reason:
      type: string
      pattern: '^(?:.|\s)*$'
      maxLength: 1073741824
      nullable: true
      example: New vessel inbound (#142)
...
    NullableTimestamp:
      allOf:
        - $ref: '#/components/schemas/Timestamp'
        - title: Nullable Timestamp
          nullable: true
    Timestamp:
      type: string
      format: date-time
      pattern: >-
        ^(-?[0-9]+)-?(1[0-2]|0[1-9])-?(3[01]|0[1-9]|[12][0-9])T(2[0-3]|[01][0-9]):?([0-5][0-9]):?([0-5][0-9])Z$
      maxLength: 22
      example: 2022-05-11T16:32:35Z
...
    Schedule:
      title: Schedule
      description: >-
        A Schedule object has the following attributes:


        + vessel: `1` (integer) - ID of the Vessel.
        
        + manual: `false` (boolean) - True if the Vessel should be planned at the supplied Berth and timestamp or false if it should be planned by the software.
        
        + berth: `20` (integer) - ID of the Berth the Vessel should be unloaded at.
        
        + start: `2022-05-12T07:23:00Z` (timestamp) - When the Vessel should start being unloaded.
        
        + end: `2022-05-13T21:41:00Z` (timestamp) - When the Vessel is expected to end being unloaded.
        
        
        The `berth` and `start` attributes are not necessary if the `type` is `automatic`.
        The `end` attribute will be calculated by the server.
      type: object
      properties:
        vessel:
          $ref: '#/components/schemas/VesselID'
        manual:
          type: boolean
        berth:
          $ref: '#/components/schemas/BerthID'
        start:
          $ref: '#/components/schemas/Timestamp'
        end:
          $ref: '#/components/schemas/Timestamp'
      additionalProperties: false
      required:
        - vessel
        - manual
        - berth
        - start
      example:
        vessel: 1
        manual: true
        berth: 30
        start: 2022-05-11T16:32:00Z
        end: 2022-05-11T21:26:00Z
...
    VesselID:
      title: Vessel ID
      description: >-
        An ID of a Vessel. Vessel IDs are unique, and are generated by the
        database.
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
...
    Vessel:
      title: Vessel
      description: >-
        A Vessel object has the following attributes:


        + name: `BigBoatHan` (string) - The name of the Vessel.

        + arrival: `2022-05-11T13:10:00Z` (timestamp) - When the Vessel is expected to arrive.

        + deadline: `2022-05-12T09:54:00Z` (timestamp) - When the Vessel must be completely unloaded.

        + containers: `5` (integer) - The amount of containers on the Vessel to be unloaded.

        + cost_per_hour: `2.6` (decimal) - The cost per hour from arrival till fully unloaded.

        + destination: `3` (integer) - The terminal where the Vessel should be unloaded.

        + length: `30` (integer) - The length of the Vessel in metres.

        + width: `10` (integer) - The width of the Vessel in metres.

        + depth: `20` (integer) - The depth of the Vessel in metres.
      type: object
      properties:
        name:
          type: string
          pattern: '^(?:.|\s)+$'
          maxLength: 64
        arrival:
          $ref: '#/components/schemas/Timestamp'
        deadline:
          $ref: '#/components/schemas/NullableTimestamp'
        containers:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        cost_per_hour:
          type: number
          format: real
          minimum: 0
          maximum: 1e+37
        destination:
          $ref: '#/components/schemas/TerminalID'
        length:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        width:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        depth:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
      additionalProperties: false
      required:
        - name
        - arrival
        - containers
        - cost_per_hour
        - destination
        - length
        - width
        - depth
      example:
        name: Vessel 1
        arrival: 2022-05-11T13:56:00Z
        deadline: 2022-05-12T18:13:00Z
        containers: 30
        cost_per_hour: 2.6
        destination: 4
        length: 30
        width: 10
        depth: 30
...
    BerthID:
      title: Berth ID
      description: >-
        An ID of a Berth. Berth IDs are unique, and are generated by the
        database.
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
...
    TerminalID:
      title: Terminal ID
      description: >-
        An ID of a Terminal. Terminal IDs are unique, and are generated by the
        database.
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
...
    Email:
      title: Email
      description: An email address for an Employee
      type: string
      format: email
      pattern: '^[A-Za-z0-9._+%-]+@[A-Za-z0-9.-]+[.][A-Za-z]+$'
      maxLength: 64
      example: jdoe@tormails.com
...
  responses:
    NoContent:
      description: 204 No Content
    NotFound:
      description: 404 Not Found
    Unauthorized:
      description: 401 Unauthorized
    Forbidden:
      description: 403 Forbidden
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: You must generate a JWT through the /token/obtain endpoint and then pass it as a Bearer token in the Authorization header.
...